chore: sd-interactive docs
