+++ packages/components/src/styles/interactive/interactive.stories.ts
@@ -8,51 +8,23 @@ const { overrideArgs } = storybookHelpers('sd-interactive');
 const { generateTemplate } = storybookTemplate('sd-interactive');
 
 /**
- * Generates basic styles for interactive elements. This should always be used in combination with `a` or `button` elements.
- * The `sd-interactive--reset` class resets the default browser styles of e. g. a button and was applied to the `button` element in the examples below.
+ * Used in interactive elements as a "quartery" button that has no paddings and no background. Its interaction states mirror those of sd-link’s.
+ *
+ *  * **Related templates**:
+ * - [Interactive](?path=/docs/templates-interactive--docs)
  */
 
 export default {
   title: 'Styles/sd-interactive',
+  tags: ['!dev'],
   component: 'sd-interactive',
-  parameters: {
-    ...parameters
-  },
+  parameters: parameters,
   args: overrideArgs([
     { type: 'slot', name: 'default', value: 'Lorem Ipsum' },
     { type: 'attribute', name: 'sd-interactive--reset', value: true }
   ]),
-  argTypes,
-  decorators: [
-    (story: any) =>
-      html`<style>
-          button.sd-interactive:not(.sd-interactive--reset) {
-            background-color: rgb(239, 239, 239);
-            border: 1px solid rgb(204, 204, 204);
-            border-radius: 4px;
-            color: rgb(51, 51, 51);
-            cursor: pointer;
-            font-family: 'Roboto', sans-serif;
-            font-size: 14px;
-            font-weight: 500;
-            line-height: 20px;
-            padding: 8px 16px;
-          }
-          button.sd-interactive:not(.sd-interactive--reset):hover {
-            background-color: rgb(230, 230, 230);
-            border-color: rgb(204, 204, 204);
-            color: rgb(51, 51, 51) !important;
-          }
-          button.sd-interactive:not(.sd-interactive--reset):active {
-            background-color: rgb(204, 204, 204);
-          }</style
-        >${story()}`
-  ]
+  argTypes
 };
-
-/**
- * Default: This shows sd-interactive in its default state.
- */
 
 export const Default = {
   render: (args: any) => {
@@ -64,88 +36,43 @@ export const Default = {
 };
 
 /**
- * Use the `disabled` class to make an interactive element look disabled. This works as well when setting an `disabled` attribute on the element.
- */
-
-export const Disabled = {
-  parameters: { controls: { exclude: ['sd-interactive--disabled', 'sd-interactive--reset'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: [{ type: 'attribute', name: 'sd-interactive--disabled', values: [false, true] }]
-      },
-      options: { templateContent: '<button class="%CLASSES%">%SLOT%</button>' },
-      constants: { type: 'attribute', name: 'sd-interactive--disabled', value: true },
-      args
-    });
-  }
-};
-
-/**
- * Use the `inverted` class to invert the colors of an interactive element.
- */
-
-export const Inverted = {
-  parameters: { controls: { exclude: ['sd-interactive--inverted', 'sd-interactive--reset'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: [{ type: 'attribute', name: 'sd-interactive--inverted', values: [false, true] }]
-      },
-      options: {
-        templateContent: '<button class="%CLASSES%">%SLOT%</button>',
-        templateBackgrounds: { alternate: 'y', colors: ['white', 'rgb(var(--sd-color-primary, 0 53 142))'] }
-      },
-      constants: { type: 'attribute', name: 'sd-interactive--inverted', value: true },
-      args
-    });
-  }
-};
-
-/**
- * Use the `reset` class to reset the default browser styles of e. g. a button.
+ * Use the `sd-interactive--reset` class to reset the default browser styles of e. g. a button.
  */
 
 export const Reset = {
-  parameters: { controls: { exclude: ['sd-interactive--reset'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: [{ type: 'attribute', name: 'sd-interactive--reset', values: [false, true] }]
-      },
-      options: { templateContent: '<button class="%CLASSES%">%SLOT%</button>' },
-      constants: { type: 'attribute', name: 'sd-interactive--reset', value: true },
-      args
-    });
-  }
+  render: () => html`
+    <div class="flex flex-row gap-12">
+      <button type="button" class="sd-interactive">Lorem Ipsum</button>
+      <button class="sd-interactive sd-interactive--reset">Lorem Ipsum</button>
+    </div>
+  `
 };
 
 /**
- * Examples: This shows how sd-interactive looks in different contexts.
+ * Use the `sd-interactive--inverted` class when displayed on primary background.
+ *
+ * On darker backgrounds text/white inverted colour will substitute primary.
  */
 
-export const Examples = {
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: {
-          type: 'slot',
-          name: 'default',
-          values: [
-            'Text',
-            '<span>Text</span>',
-            '<sd-icon library="global-resources" name="system/picture"></sd-icon>'
-          ].map(content => {
-            return {
-              value: content,
-              title: content.replace('library="global-resources" name="system/picture"', '...')
-            };
-          }),
-          title: 'content'
-        }
-      },
-      options: { templateContent: '<button class="%CLASSES%">%SLOT%</button>' },
-      args
-    });
-  }
+export const Inverted = {
+  render: () => html`
+    <div class="bg-primary p-4">
+      <button class="sd-interactive sd-interactive--inverted sd-interactive--reset">Lorem Ipsum</button>
+    </div>
+  `
+};
+
+/**
+ * Use the `sd-interactive--disabled` class to make an interactive element look disabled.
+ *
+ * This works as well when setting an `disabled` attribute on the element.
+ */
+
+export const Disabled = {
+  render: () => html`
+    <div class="flex flex-row gap-12">
+      <button class="sd-interactive sd-interactive--reset">Lorem Ipsum</button>
+      <button class="sd-interactive sd-interactive--disabled sd-interactive--reset">Lorem Ipsum</button>
+    </div>
+  `
 };
